Add prefix, Move json parser from ParseJsonImpl to ForgeTool
